Don't call $funcname if value is not set
In bw_form_field() if for some reason the $funcname returned from bw_form_function
is not set then we should not attempt to invoke it.

This fix is related to bobbingwide/oik-weightcountry-shipping/issues/6
where the Fatal error in the post editor was caused by the dummy bw_trace2()
function not returning the value passed to it.